requirements.txt
streamlit
torch
torchvision
transformers
Pillow
python-dotenv
openai>=1.0.0
